Upgrade and bug fixes
Had some issues where some if some data was not available, program would error out. For example if a patent did not have a assignee, it would error out. Implemented conditionals to only print if information was detected. Also implemented a few new functions which pulls the status and name of the patent